% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/litterDiallel.R
\name{diallelMatrixMakeAndRotate}
\alias{diallelMatrixMakeAndRotate}
\title{diallelMatrixMakeAndRotate}
\usage{
diallelMatrixMakeAndRotate(data, dam.col.name, sire.col.name,
  batch.col.name = NULL, batch.1.col.name = NULL, n.strains = 8, ...)
}
\arguments{
\item{data}{data frame}

\item{dam.col.name}{dam column name}

\item{sire.col.name}{sire column name}

\item{batch.col.name}{name of batch/random effect column}

\item{batch.1.col.name}{name of additional batch/random effect column}

\item{...}{additional arguments}
}
\value{
returns diallel incidence matrices, rotated
}
\description{
Make design matrices for diallel, rotate to n-1 space.
}
\examples{
## not run
}
