% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/litterDiallel.R
\name{diallelMatrixMakerSexed}
\alias{diallelMatrixMakerSexed}
\title{diallelMatrixMakerSexed}
\usage{
diallelMatrixMakerSexed(
  data,
  dam.col.name,
  sire.col.name,
  batch.col.name = NULL,
  batch.1.col.name = NULL,
  strains = c("AJ", "B6", "129", "NOD", "NZO", "CAST", "PWK", "WSB"),
  sex.col.name,
  female.sex.label,
  ...
)
}
\arguments{
\item{data}{data frame}

\item{dam.col.name}{dam column name}

\item{sire.col.name}{sire column name}

\item{batch.col.name}{name of batch/random effect column}

\item{batch.1.col.name}{name of additional batch/random effect column}

\item{strains}{character vector indicating expected strain names}

\item{sex.col.name}{the column name for sex}

\item{female.sex.label}{the character string indicating female sex}

\item{...}{additional arguments}
}
\value{
returns diallel incidence matrices
}
\description{
Make design matrices for diallel, fulls model
}
\examples{
## not run
}
