% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/litterDiallel.R
\name{ifow}
\alias{ifow}
\title{ifow}
\usage{
ifow(test, yes, no)
}
\arguments{
\item{test}{test}

\item{yes}{specify yes}

\item{no}{specify no}
}
\value{
returns yes if test is true, no if false
}
\description{
if, otherwise
}
\examples{
## not run
}
