% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/litterDiallel.R
\docType{package}
\name{litterDiallel}
\alias{litterDiallel}
\title{litterDiallel: A package for reproducing the analysis used in the litter diallel manuscript.}
\description{
A package for reproducing the analysis used in the litter diallel manuscript.
}
