% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/litterDiallel.R
\name{makeRotationMatrix}
\alias{makeRotationMatrix}
\title{makeRotationMatrix: Make a rotation matrix}
\usage{
makeRotationMatrix(X, n, ...)
}
\arguments{
\item{X}{original design matrix}

\item{n}{number of original columns of design matrix}

\item{...}{additional arguments}
}
\value{
generates rotation matrix to reduce dimensions for better sampling
}
\description{
Turn an n-column design matrix into an n-1, sum to 0, design matrix
             while maintaining independence. 
             See Appendix (Lenarcic, 2012, Genetics; Crowley, 2014, Genetics)
}
\examples{
## not run
}
