% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/litterDiallel.R
\name{mcmc.stack}
\alias{mcmc.stack}
\title{mcmc.stack}
\usage{
mcmc.stack(coda.object, ...)
}
\arguments{
\item{coda.object}{coda object with multiple chains}

\item{...}{additional arguments}
}
\value{
returns stacked mcmc
}
\description{
stack multiple mcmc chains into one mcmc object
}
\examples{
## not run
}
