% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/litterDiallel.R
\name{plot.hpd}
\alias{plot.hpd}
\title{plot.hpd: Plot highest posterior density intervals (from BayesDiallel package, Will Valdar and Alan Lenarcic)}
\usage{
\method{plot}{hpd}(coda.object, wanted = varnames(coda.object),
  prob.wide = 0.95, prob.narrow = 0.5, xlab = "HPD interval",
  names = NULL, type = "p", name.margin = 6.1, plt.left = NULL,
  plt.right = NULL, plt.bottom = NULL, plt.title = NULL, ylab = "",
  name.line = 3.9, main = "", main.line = 2, ...)
}
\arguments{
\item{coda.object}{coda object}

\item{wanted}{variable names for coda object}

\item{prob.wide}{outer width of posterior probability}

\item{prob.narrow}{inner width of posterior probability}

\item{xlab}{x-axis label}

\item{names}{names}

\item{type}{type of plot}

\item{name.margin}{margin for name}

\item{plt.left}{left of plot margin}

\item{plt.right}{right of plot margin}

\item{plt.bottom}{bottom of plot margin}

\item{plt.title}{title of plot margin}

\item{ylab}{y-axis label}

\item{name.line}{line where names should be printed}

\item{main}{text of main title}

\item{main.line}{line where main title should be printed}

\item{...}{additional arguments}
}
\value{
returns HPD plot
}
\description{
Plot HPD intervals.
}
\examples{
## not run
}
