% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/litterDiallel.R
\name{plot_ci}
\alias{plot_ci}
\title{plot_ci}
\usage{
plot_ci(
  midvals,
  narrow.intervals,
  wide.intervals,
  names = 1:length(midvals),
  add = FALSE,
  main = "",
  main.line = 2,
  xlab = "Estimate",
  xlab.line = 2.5,
  xlim = NULL,
  ylab = "",
  yaxis = TRUE,
  ylim = c(0, length(midvals)),
  name.line = 4,
  pch.midvals = 19,
  col = "black",
  col.midvals = col,
  cex.labels = 1,
  type = "p",
  name.margin = 6.1,
  title.margin = 4.1,
  title.line = 3.5,
  bottom.margin = 5.1,
  bottom.line = 4.5,
  right.margin = 2.1,
  right.line = 1.5,
  mar = sides(left = name.margin, bottom = bottom.margin, top = title.margin, right =
    right.margin),
  mar.update = sides(),
  before.data = function() { },
  plt.left = NULL,
  plt.right = NULL,
  plt.bottom = NULL,
  plt.title = NULL,
  ...
)
}
\arguments{
\item{midvals}{midvals}

\item{narrow.intervals}{narrow.intervals}

\item{wide.intervals}{wide.intervals}

\item{names}{names}

\item{add}{add}

\item{main}{main}

\item{main.line}{main.line}

\item{xlab}{xlab}

\item{xlab.line}{xlab.line}

\item{xlim}{xlim}

\item{ylab}{ylab}

\item{yaxis}{yaxis}

\item{ylim}{ylim}

\item{name.line}{name.line}

\item{pch.midvals}{pch.midvals}

\item{col}{col}

\item{col.midvals}{col.midvals}

\item{cex.labels}{cex.labels}

\item{type}{type}

\item{name.margin}{name.margin}

\item{title.margin}{title.margin}

\item{title.line}{title.line}

\item{bottom.margin}{bottom.margin}

\item{bottom.line}{bottom.line}

\item{right.margin}{right.margin}

\item{right.line}{right.line}

\item{mar}{mar}

\item{mar.update}{mar.update}

\item{before.data}{before.data}

\item{plt.left}{plt.left}

\item{plt.right}{plt.right}

\item{plt.bottom}{plt.bottom}

\item{plt.title}{plt.title}

\item{...}{additional arguments}
}
\value{
returns plotted caterpillar plot with confidence intervals
}
\description{
stack multiple mcmc chains into one mcmc object
}
\examples{
## not run
}
