% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/litterDiallel.R
\name{sides}
\alias{sides}
\title{sides}
\usage{
sides(
  default = NA,
  bottom = default,
  left = default,
  top = default,
  right = default
)
}
\arguments{
\item{default}{specify overall default}

\item{bottom}{specify bottom}

\item{left}{specify left}

\item{top}{specify top}

\item{right}{specify default}
}
\value{
returns named side arguments
}
\description{
define sides of a plot
}
\examples{
## not run
}
