% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/litterDiallel.R
\name{update_sides}
\alias{update_sides}
\title{update_sides}
\usage{
update_sides(old = graphics::par("mar"), new = rep(NA, 4))
}
\arguments{
\item{old}{defaults to par("mar")}

\item{new}{defaults to 4 NA's}
}
\value{
returns updated sides
}
\description{
update sides of a plot
}
\examples{
## not run
}
