% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/litterDiallel.R
\name{incidence.matrix}
\alias{incidence.matrix}
\title{incidence.matrix: Make an incidence matrix}
\usage{
incidence.matrix(fact, ...)
}
\arguments{
\item{fact}{factor}

\item{...}{additional arguments}
}
\value{
generates an incidence matrix
}
\description{
Convert a factor into an incidence matrix. (From the WVmisc package, Will Valdar).
}
\examples{
## not run
}
